Use /bin/true as a no-op.
We need a no-op if `-lcriterion` is found, because we don't want the
default behavior of adding it to `LDADD`. Just doing a `/bin/true`
shell command is a little better than printing a result of "yes",
which actually just shows up as an extra line with "yes" in it
during a run of `configure`.