Ensure game can start with various missing files.
This means we don't have to install the zero-length version that
originally was distributed in the `lib/` directory for several
of them. The game will create zero-length versions in-place as
needed. Some of the other zero-length files in `lib/` do not
appear to be needed, as well.

This addresses the remainder of #26.