Fix parsing bug for numbered items.
This was crashing on trying to parse a numbered item like
"2.hammer". The issue was a `strcpy` where the source and
destination strings overlapped, for which "behavior is
undefined" (so I guess an `abort` is a fine thing for a
library to do). This is rewritten to use `memmove` instead.

In addition, since `get_number` never actually modifies the
`char *` that gets passed in as a parameter, there's no need
for the extra level of indirection, and we can simplify its
prototype.