Change appropriate SPRINTs to snprintf
gcc 4.8 (though not gcc5) objects to the macro def of SPRINTF when the
dest is a pointer instead of an array. It is trying to be helpful and
the warning is triggered b/c the sizeof check is done at runtime, not
compile time. So just explicitly fix those cases to use snprintf. I
wasn't crazy about that macro test anyway.