Separate out temporary index from type-specific index creation.
It turns out that not the right thing was going on here. The
intent was that the initial call to `generate_indices` would
populate `top_of_mobt` or `top_of_objt`, which would then be
passed as the second argument to either `make_mob_indices` or
`make_obj_indices`. This works locally for me on Mac OS X with
the Apple LLVM-based `gcc`.

However, when running on the prod Linux machine when compiled
with `gcc` 4.8.3, the order of operations doesn't come out
right. In this case, the _original_ value for `top_of_mobt` or
`top_of_objt` gets passed in: `0`, and hence the new indices
don't get created correctly.

By un-lining the call to `generate_indices` we can make sure
it is unambiguous what order we want these to happen in.