Fix learning at ninja master.
Special procs return TRUE if they handle a command; in this case
a successful practice at the ninja master was not returning it
from the right place, and it was falling through to the default
FALSE, which is why it got handled by the `do_practice` command
in addition.

Addresses #81.