Tidy up version numbering.
We (ok, I) have been doing git tags for releases like "v0.1.9"
but this results in the use of that version string downstream.
For example, `make dist` will produce `sillymud-v0.1.9.tar.gz`
instead of the desired `sillymud-0.1.9.tar.gz`. This change
just strips the leading `v` off the tag name to use for the
_autotools_ version.

However, I would also be ok with going back through to update
the git tags to be the version numbers we want (`0.1.9`) under
the assumption that probably no one would care. Putting this
PR to collect opinions on the best path forward.