Fixed some broken records in the mob file.
Some of the records in `tinyworld.mob` were malformed. It's hard
to know exactly what the intent was with some of these, so the
fixes may or may not be in line with what the original
implementors wanted. However, they are now syntactically correct,
and I did my best to line them up with other similar-seeming mobs
nearby in the file.