Fix autoconf detection of Criterion library
The existing `AC_CHECK_LIB` test was trying to find a symbol that
was not, apparently, exported by the library. Having used the
`strings` command to examine the lib revealed a function we can
use successfully for the test.

In addition, the `AC_CHECK_HEADERS` was only partially finding
the Criterion header files; turns out `-I` including them in *both*
`CFLAGS` and `CPPFLAGS` (even though we don't have any C++ in the
project) gets the test to pass properly. Updated `README.md` to
document this.