Restore scripts to working order
I noticed when booting the game just now that it was complaining
about not being able to load script files. Turns out, the source
tarball this repo was based on did not include any of the scripts!
Fortunately, the Interwebz had copies of them, so the scripts are
back in action!

(Also, kudos to the prior devs, who were handling the missing
script files gracefully without crashing!)